Add command line flag for multiprocessing

+++ bitwise_challenge_2022_2/__main__.py
@@ -18,11 +18,15 @@ def main(argv: list[str] = None):
     Args:
         argv (list[str]): List of command line arguments. Defaults to None.
     """
-    n_threads = multiprocessing.cpu_count()
-    pool = multiprocessing.Pool(n_threads, _init_worker)
-    try:
-        parsed_args = _parse_args(argv)
+    parsed_args = _parse_args(argv)
 
+    if parsed_args.multiprocessing:
+        n_threads = multiprocessing.cpu_count()
+        pool = multiprocessing.Pool(n_threads, _init_worker)
+    else:
+        pool = None
+
+    try:
         network_json = Path(__file__).parent / "koodipahkina-data.json"
         res = optimize(
             network_json,
@@ -41,12 +45,14 @@ def main(argv: list[str] = None):
 
     except KeyboardInterrupt:
         print("\nInterrupted, exiting")
-        pool.terminate()
-        pool.join()
+        if pool is not None:
+            pool.terminate()
+            pool.join()
         sys.exit(1)
     else:
-        pool.close()
-        pool.join()
+        if pool is not None:
+            pool.close()
+            pool.join()
 
 
 def _init_worker():
@@ -85,6 +91,11 @@ def _parse_args(args: list[str]) -> argparse.Namespace:
         help="Location for saving intermediate x values for population. Default: opt_X_latest.npy",
     )
     parser.add_argument("--resume", action="store_true", help="Resume from xfile")
+    parser.add_argument(
+        "--multiprocessing",
+        action="store_true",
+        help="Use multiprocessing for problem evaluation",
+    )
     parser.add_argument("--quiet", "-q", action="store_true", help="Suppress output")
 
     return parser.parse_args(args)
